+++ web/src/components/MomentaryChat.tsx
@@ -4,7 +4,7 @@ import DOMPurify from "dompurify";
 import "./MomentaryChat.css";
 import { useMomentaryChat, ChatItem } from "../hooks/useMomentaryChat";
 import { EmojiPicker } from "../utils/emoji";
-import { WysiwygEditor } from "./WysiwygEditor";
+import { WysiwygEditor } from "./reusable/WysiwygEditor";
 import { useNotification } from "../hooks/useNotification";
 
 const MAX_CHAT_TEXT_SIZE = 1 * 1024 * 1024;
