+++ web/src/components/reusable/FaceCard.tsx
@@ -1,9 +1,9 @@
 import { memo } from "react";
 
 import "./FaceCard.css";
-import { useFaceImageObsoleted } from "../hooks/useFaceImages";
-import { BLANK_IMAGE } from "../media/imagePresets";
-import { isEmoji } from "../utils/emoji";
+import { useFaceImageObsoleted } from "../../hooks/useFaceImages";
+import { BLANK_IMAGE } from "../../media/imagePresets";
+import { isEmoji } from "../../utils/emoji";
 
 export const FaceCard = memo<{
   image?: string;
